Refactor age range into date of birth
